+++ src/command/sample_encode.rs
@@ -168,6 +168,7 @@ pub async fn run(
             input_len,
             full_pass,
             &enc_args,
+            &vmaf,
         )
         .await
         {